requirements.txt
streamlit>=1.28.0
ultralytics>=8.0.196
Pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0